+++ examples/data-sources/agile_logical_switch/data-source.tf
@@ -1,0 +1,3 @@
+data "agile_logical_switch" "example" {
+  name = "example"
+}